working/requirements.txt
tqdm
pycocotools
tensorboard
tensorboardX
pyyaml
webcolors
matplotlib
numpy==1.21.3
opencv-python
opencv-contrib-python
Pillow
PyMySQL==1.0.2
torch==1.9.0
torchvision==0.10.0